requirements.txt
python-telegram-bot==22
python-dotenv==1.1.0
requests==2.32.3
aiohttp==3.11.14